+++ NEWS.md
@@ -65,6 +65,18 @@ ClimaAnalysis.add_file!(catalog, "precip.nc", "precip" => "pr")
 pr_var = get(catalog, "pr")
 ```
 
+## Mask aware flatten
+
+With this release, you can pass in a mask to `flatten` to do a mask aware
+flatten, where all the values corresponding to zeros in the mask are dropped.
+Any value of the `OutputVar` whose coordinates correspond to zeros on the mask are
+excluded from the flattened data.
+
+```julia
+ocean_mask = ClimaAnalysis.generate_ocean_mask(NaN, 1.0; threshold = 0.5)
+flat_masked_var = ClimaAnalysis.flatten(var, mask = ocean_mask)
+```
+
 v0.5.18
 -------
 This release introduces the following features and bug fixes